Add "use-system" feature to specify that external metis is required
The code now assumes "vendored" by default, which builds and statically links to METIS and GKLib. This is a shift from the previous setup, which expected an existing METIS install and dynamically linked to it. The support for use of existing METIS install is preserved, but now under a new feature - "use-system".